Refactor GridPersistenceModel and ZoneGridPersistenceModel into standalone util functions
